Moving averaging trials under .zaccum package
