+++ grvx_mvn/grvx_zpol/src/main/scala/fun/gravax/distrib/struct/BinTreeStuff.scala
@@ -28,6 +28,7 @@ trait BinTreeLoader extends KnowsBinItem with KnowsBinTupTupTypes {
 		// So to "fetch any number" we would want a Stream.
 		// But perhaps for now we will be content to just fetch blocks of child nodes.
 
-
 	}
+
+	def mkBinDataRecAndNode
 }
